Move code to proper locations.
