+++ api/clients/request_signer.go
@@ -2,6 +2,7 @@ package clients
 
 import (
 	"context"
+	"crypto/ecdsa"
 	"fmt"
 	grpc "github.com/Layr-Labs/eigenda/api/grpc/node/v2"
 	"github.com/Layr-Labs/eigenda/node/auth"
@@ -21,24 +22,39 @@ var _ RequestSigner = &requestSigner{}
 
 type requestSigner struct {
 	keyID      string
+	publicKey  *ecdsa.PublicKey
 	keyManager *kms.Client
 }
 
 // NewRequestSigner creates a new RequestSigner.
 func NewRequestSigner(
+	ctx context.Context,
 	region string,
 	endpoint string,
-	keyID string) RequestSigner {
+	keyID string) (RequestSigner, error) {
 
 	keyManager := kms.New(kms.Options{
 		Region:       region,
 		BaseEndpoint: aws.String(endpoint),
 	})
 
+	getPublicKeyOutput, err := keyManager.GetPublicKey(ctx, &kms.GetPublicKeyInput{
+		KeyId: aws.String(keyID),
+	})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get public key: %w", err)
+	}
+
+	publicKey, err := auth.ParseKMSPublicKey(getPublicKeyOutput.PublicKey)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse public key: %w", err)
+	}
+
 	return &requestSigner{
 		keyID:      keyID,
+		publicKey:  publicKey,
 		keyManager: keyManager,
-	}
+	}, nil
 }
 
 func (s *requestSigner) SignStoreChunksRequest(ctx context.Context, request *grpc.StoreChunksRequest) ([]byte, error) {
@@ -57,5 +73,10 @@ func (s *requestSigner) SignStoreChunksRequest(ctx context.Context, request *grp
 		return nil, fmt.Errorf("failed to sign request: %w", err)
 	}
 
-	return signOutput.Signature, nil
+	signature, err := auth.ParseKMSSignature(s.publicKey, hash, signOutput.Signature)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse signature: %w", err)
+	}
+
+	return signature, nil
 }
